Error Fix
Adding a delay of 10000 miliseconds after setting the values to the backend spreadsheets, so the backend calculates everything during the 10 seconds and just when it is finished the script continues to get the values. This was needed to avoid an error that was appearing when the script tried to pull the values from the backend spreadsheet, but it still hasn't finished calculating the costs. Also removes the space between the % and the number it shows.